sodan

+++ TestProject/src/Interfaces/IPythagoras.java
@@ -2,6 +2,6 @@ package Interfaces;
 
 public interface IPythagoras {
 	
-	float calcHypotenuse(float a, float b);
+	double calcHypotenuse(float a, float b);
 
 }
